refactor: 딜러 결과 계산하는 메서드 위치 딜러로 변경
